fixed ensemble handling in training
