+++ scripts/reach2d_exps/bc_oracle_mlp.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 python src/main.py \
-	--exp_name dec24/oracle_reach2d_mlp_test_refactor \
+	--exp_name dec25/oracle_reach2d_mlp_test_refactor \
     --data_path ./data/scripted_oracle_reach2d.pkl \
     --environment Reach2D \
     --method BC \
