+++ scripts/reach2d_exps/dagger_oracle_mlp.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 python src/main.py \
-	--exp_name dec24/dagger_reach2d_mlp_test_refactor \
+	--exp_name dec25/dagger_reach2d_mlp_test_refactor \
     --data_path ./data/scripted_oracle_reach2d.pkl \
     --environment Reach2D \
     --method Dagger \
