+++ scripts/reach2d_exps/ensemble_dagger_oracle_mlp.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 python src/main.py \
-	--exp_name dec25/test_bc_oracle_reach2d_refactor \
+	--exp_name dec25/ensemble_dagger_reach2d_mlp_test_refactor \
     --data_path ./data/scripted_oracle_reach2d.pkl \
     --environment Reach2D \
-    --method BC \
-    --arch LinearModel \
-    --num_models 1 \
+    --method Dagger \
+    --arch MLP \
+	--hidden_size 20 \
+    --num_models 3 \
     --seed 4 \
 	--overwrite
