+++ scripts/reach2d_exps/eval_bc_oracle_linear.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 python src/main.py \
-    --exp_name dec24/test_bc_oracle_reach2d_refactor_eval \
+    --exp_name dec25/test_bc_oracle_reach2d_refactor_eval \
 	--eval_only \
-	--model_path ./out/dec24/test_bc_oracle_reach2d_refactor/model_4.pt\
+	--model_path ./out/dec25/test_bc_oracle_reach2d_refactor/model_4.pt\
     --environment Reach2D \
     --method BC \
     --arch LinearModel \
