+++ scripts/reach2d_exps/eval_dagger_oracle_mlp.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 python src/main.py \
-	--exp_name dec24/dagger_reach2d_mlp_test_refactor_eval \
+	--exp_name dec25/dagger_reach2d_mlp_test_refactor_eval \
 	--eval_only \
-	--model_path ./out/dec24/dagger_reach2d_mlp_test_refactor/model_4.pt\
+	--model_path ./out/dec25/dagger_reach2d_mlp_test_refactor/model_4.pt\
     --environment Reach2D \
     --method Dagger \
     --arch MLP \
