+++ src/util/__init__.py
@@ -1,2 +1,2 @@
 from util.constant import *
-from util.util import get_model_type_and_kwargs
+from util.util import get_model_type_and_kwargs, setup_robosuite
