+++ src/util/constant.py
@@ -1,4 +1,12 @@
-ACT_MAGNITUDE = 0.1
-MAX_REACH2D_TRAJ_LEN = 100
-MAX_PICKPLACE_TRAJ_LEN = 175
+# Dataset buffer size
+MAX_BUFFER_SIZE = int(3e4)
+
+# PickPlace constants
+PICKPLACE_MAX_TRAJ_LEN = 175
+
+# Reach2D constants
+REACH2D_ACT_MAGNITUDE = 0.1
+REACH2D_MAX_TRAJ_LEN = 100
+REACH2D_ACT_DIM = 2
+REACH2D_OBS_DIM = 4
 REACH2D_SUCCESS_THRESH = 0.1
